requirements.txt
streamlit
pandas
scikit-learn
requests
joblib
python-dotenv
